requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
sqlparse==0.4.3
tzdata==2022.4
pytz==2022.6
django-widget-tweaks==1.4.12